code refactor for logged in user view
